requirements.txt
requests
beautifulsoup4
PyYAML
lxml